requirements.txt
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
Django==2.2.24
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
idna==2.10
importlib-metadata==2.1.1
iniconfig==1.1.1
packaging==20.9
pathlib2==2.3.6
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.1.2
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
zipp==1.2.0
